Refactoring & code mantain
